requirements.txt
dspy-ai
pydantic
pydantic-settings
pyyaml
google-generativeai
python-dotenv
tenacity
markdown
beautifulsoup4
rich
